Initial Commit
Uses the first webcam it can find and then opens a window where you can view the pixel displacement of the target. Minds that the bounding box is off by a factor of +3px from testing and all values for distance are placeholder and dependent on the camera which you use. Also, there is a bug where on Ubuntu you can't view the video unless there is a one-milliseconds wait at the end of the while loop.